refactor(runtime): delete test-only session projection
