+++ src/test-kernel/agent/sessionProgressTestUtils.ts
@@ -1,10 +1,202 @@
+import type { AgentEvent } from '@agent/trace';
 import type { AgentRuntimeHost } from '@agent/runtime/AgentRuntimeHost';
-import type { ProjectedProgressEvent } from '@agent/runtime/sessionProgressEventProjection';
+import type {
+  ProgressEvent,
+  ProgressEventPayloads,
+} from '@agent/runtime/hostProgressEvents';
+import { fromRunFactDomainKey } from '@agent/runtime/runFactEvents';
+import { toUpdateStreamUsagePayload } from '@agent/runtime/runFactUsage';
+import type {
+  SessionEventHub,
+  SessionFact,
+} from '@agent/runtime/SessionEventHub';
+import { agentConfigToTaskState } from '@agent/utils/agentConfigToTaskState';
 import {
-  projectSessionFactToProgressEvent,
-  subscribeRunFactsAsProgressEvents,
-} from '@agent/runtime/sessionProgressEventProjection';
-import type { SessionEventHub } from '@agent/runtime/SessionEventHub';
+  ConversationProgressSchema,
+  UpdatePlanPayloadSchema,
+  UpdateTodosPayloadSchema,
+  type StreamTabId,
+} from '@shared/schemas';
+import { isObject } from '@utils/core';
+
+type ProjectedProgressEvent = {
+  [K in ProgressEvent]: {
+    readonly event: K;
+    readonly payload: ProgressEventPayloads[K];
+  };
+}[ProgressEvent];
+
+function projectSessionFactToProgressEvent(
+  fact: SessionFact,
+): ProjectedProgressEvent | undefined {
+  switch (fact.type) {
+    case 'goalStateChanged':
+      return { event: 'goalStateChanged', payload: fact.payload };
+    case 'inquiryThreadUpdated':
+      return { event: 'inquiryThreadUpdated', payload: fact.payload };
+    case 'clearMissingOutputs':
+      return { event: 'clearMissingOutputs', payload: fact.payload };
+    case 'updateQueuedFollowUps':
+      return { event: 'updateQueuedFollowUps', payload: fact.payload };
+    case 'followUpSent':
+      return undefined;
+    case 'setActiveStream':
+      return { event: 'setActiveStream', payload: fact.payload };
+    case 'updateStreamDescription':
+      return { event: 'updateStreamDescription', payload: fact.payload };
+    case 'updateStreamStatus':
+      return { event: 'updateStreamStatus', payload: fact.payload };
+    case 'setParentStream':
+      return { event: 'setParentStream', payload: fact.payload };
+    case 'removeStream':
+      return { event: 'removeStream', payload: fact.payload };
+  }
+}
+
+function projectRunFactToProgressEvent(
+  streamId: StreamTabId,
+  event: AgentEvent,
+): ProjectedProgressEvent | undefined {
+  if (event.type === 'usage') {
+    const payload = toUpdateStreamUsagePayload(event.data, streamId);
+    return payload ? { event: 'updateStreamUsage', payload } : undefined;
+  }
+
+  if (event.type === 'run.config') {
+    return {
+      event: 'setTaskState',
+      payload: {
+        streamId: event.streamId,
+        executionId: event.executionId,
+        taskState: agentConfigToTaskState(event.config),
+      },
+    };
+  }
+
+  if (event.type === 'status') {
+    return {
+      event: 'updateStreamStatus',
+      payload: {
+        streamId: event.streamId,
+        status: event.phase,
+        cause: event.cause,
+        ...(event.previousPhase ? { previousStatus: event.previousPhase } : {}),
+        ...(event.substate ? { substate: event.substate } : {}),
+      },
+    };
+  }
+
+  if (event.type === 'domain') {
+    if (event.key === 'conversationProgress') {
+      const progress = ConversationProgressSchema.safeParse(event.data);
+      if (!progress.success) return undefined;
+      return {
+        event: 'updateConversationProgress',
+        payload: {
+          streamId,
+          progress: progress.data,
+        },
+      };
+    }
+
+    const factName = fromRunFactDomainKey(event.key);
+    if (!factName || !isObject(event.data)) return undefined;
+    if (factName === 'updateTodos') {
+      const payload = UpdateTodosPayloadSchema.safeParse(event.data);
+      return payload.success
+        ? { event: 'updateTodos', payload: payload.data }
+        : undefined;
+    }
+    if (factName === 'updatePlan') {
+      const payload = UpdatePlanPayloadSchema.safeParse(event.data);
+      return payload.success
+        ? { event: 'updatePlan', payload: payload.data }
+        : undefined;
+    }
+    return {
+      event: factName,
+      payload: event.data as unknown as ProgressEventPayloads[typeof factName],
+    } as ProjectedProgressEvent;
+  }
+
+  if (event.type === 'stage.start') {
+    if (event.kind !== 'round') return undefined;
+    return {
+      event: 'updateRoundStage',
+      payload: {
+        streamId,
+        roundStage: {
+          index: event.index ?? 0,
+          ...(event.total !== undefined && event.total > 0
+            ? { total: event.total }
+            : {}),
+        },
+      },
+    };
+  }
+
+  if (event.type === 'child.activity') {
+    if (event.kind === 'subagents') {
+      return {
+        event: 'updateActiveSubagents',
+        payload: {
+          parentStreamId: event.parentStreamId,
+          children: [...event.children],
+        },
+      };
+    }
+    if (event.kind === 'processes') {
+      return {
+        event: 'updateActiveProcesses',
+        payload: {
+          parentStreamId: event.parentStreamId,
+          processes: [...event.processes],
+        },
+      };
+    }
+  }
+
+  if (event.type === 'process.output') {
+    return {
+      event: 'updateProcessOutput',
+      payload: {
+        parentStreamId: event.parentStreamId,
+        executionId: event.executionId,
+        stdout: event.stdout,
+        stderr: event.stderr,
+      },
+    };
+  }
+
+  return undefined;
+}
+
+const RUN_FACT_PROGRESS_EVENT_TYPES: readonly AgentEvent['type'][] = [
+  'domain',
+  'run.config',
+  'usage',
+  'status',
+  'stage.start',
+  'child.activity',
+  'process.output',
+];
+
+function subscribeRunFactsAsProgressEventsForTest(
+  events: SessionEventHub,
+  onProjected: (projected: ProjectedProgressEvent) => void,
+): () => void {
+  return events.subscribe(
+    (sessionEvent) => {
+      if (sessionEvent.scope !== 'run') return;
+      const projected = projectRunFactToProgressEvent(
+        sessionEvent.streamId,
+        sessionEvent.event,
+      );
+      if (projected) onProjected(projected);
+    },
+    { scope: 'run', types: RUN_FACT_PROGRESS_EVENT_TYPES },
+  );
+}
 
 function emitProjectedProgressEventForTest(
   runtimeHost: AgentRuntimeHost,
@@ -36,7 +228,7 @@ export function attachSessionProgressEventProjectionForTest(
     },
     { scope: 'session' },
   );
-  const detachRunFacts = subscribeRunFactsAsProgressEvents(
+  const detachRunFacts = subscribeRunFactsAsProgressEventsForTest(
     events,
     (projected) => emitProjectedProgressEventForTest(runtimeHost, projected),
   );
